feat: generateEnabledChannelsEnum in Providers
