+++ apps/api/src/modules/providers/providers.module.ts
@@ -1,4 +1,4 @@
-import { Logger, Module } from '@nestjs/common';
+import { Module } from '@nestjs/common';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { JsendFormatter } from 'src/common/jsend-formatter';
 import { Provider } from './entities/provider.entity';
@@ -6,7 +6,7 @@ import { ProvidersService } from './providers.service';
 
 @Module({
   imports: [TypeOrmModule.forFeature([Provider])],
-  providers: [ProvidersService, JsendFormatter, Logger],
+  providers: [ProvidersService, JsendFormatter],
   exports: [TypeOrmModule],
 })
 export class ProvidersModule {}
